Huge world cup commit, with some robot.txt additions

+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Functions/fnWC_CalculateRanking_GameAge.sql
@@ -1,5 +1,5 @@
 ﻿CREATE FUNCTION dbo.fnWC_CalculateRanking_GameAge(
-	@game_code smallint,
+	@game_code INT,
 	@calculation_date datetime
 )
 RETURNS decimal(10,2)
