+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/ffl_unknown_song.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[ffl_unknown_song] (
-    [song_key]  SMALLINT      NOT NULL,
+    [song_key]  INT      NOT NULL,
     [song_name] VARCHAR (100) NULL,
     CONSTRAINT [PK_ffl_track] PRIMARY KEY CLUSTERED ([song_key] ASC)
 );
