+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/wc_penalty_shootout.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[wc_penalty_shootout] (
-    [penalty_code]  SMALLINT NOT NULL,
-    [game_code]     SMALLINT NOT NULL,
-    [player_code]   SMALLINT NOT NULL,
-    [penalty_order] SMALLINT NOT NULL,
+    [penalty_code]  INT NOT NULL,
+    [game_code]     INT NOT NULL,
+    [player_code]   INT NOT NULL,
+    [penalty_order] INT NOT NULL,
     [scored]        BIT      NOT NULL,
     CONSTRAINT [PK_wc_penalty_shootout] PRIMARY KEY CLUSTERED ([penalty_code] ASC)
 );
