+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/wc_ranking.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[wc_ranking] (
     [ranking_date] DATETIME        NOT NULL,
-    [team_code]    SMALLINT        NOT NULL,
-    [ranking]      SMALLINT        NOT NULL,
+    [team_code]    INT        NOT NULL,
+    [ranking]      INT        NOT NULL,
     [score]        DECIMAL (10, 2) NOT NULL
 );
