+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/wc_tournament_team_entry.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[wc_tournament_team_entry] (
-    [tournament_code]   SMALLINT      NOT NULL,
-    [team_code]         SMALLINT      NOT NULL,
+    [tournament_code]   INT      NOT NULL,
+    [team_code]         INT      NOT NULL,
     [round_code]        VARCHAR (10)  NOT NULL,
-    [fifa_ranking]      SMALLINT      NULL,
+    [fifa_ranking]      INT      NULL,
     [coach_name]        VARCHAR (200) NULL,
     [coach_nationality] VARCHAR (200) NULL
 );
